requirements.txt
Django==4.2.10
mysqlclient==2.2.0
django-crispy-forms==2.1
django-crispy-bootstrap5==2024.2
django-decouple==2.0.0
django-import-export==4.0.9
pymysql==1.1.1